move initial abstractions to Library-TraceCheckerUtils
